+++ lib/features/auth/auth_remote_datasource.dart
@@ -17,7 +17,7 @@ class AuthRemoteDatasource extends IAuthRemoteDatasource {
   Future<Either<Failure, Success>> mobileLogin(
       String clientId, String secretKey, String field, String uuid) async {
     try {
-      final response = await dio.post('$baseUrl/v3/auth/login',
+      final response = await dio.post('$baseUrl/v4/auth/login',
           options: Options(
             contentType: "application/json",
           ),